third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/live-edit-expected.txt

Tests live edit feature.


Running: testLiveEdit
==Source frame contents start==
function f()
{
    return "live-edited string";
}

==Source frame contents end==

Running: testLiveEditSyntaxError
==Source frame contents start==
function syntaxError()
{
    return "I'm good";
}

==Source frame contents end==

Running: testLiveEditWhenPaused
Script execution paused.
Script execution paused.
Script execution resumed.

Running: testNoCrashWhenOnlyOneFunctionOnStack
Script execution paused.
Call stack:
    0) f1 (edit-me-when-paused.js:13)
Script execution resumed.
Script execution paused.
Script execution resumed.
Script execution paused.
Script execution resumed.

Running: testBreakpointsUpdated
breakpoint at 3
  inline breakpoint at (3, 4)
  inline breakpoint at (3, 32) disabled

Running: testNoCrashWhenLiveEditOnBreakpoint
Script execution paused.
Script execution resumed.
